refactor: organize components
move components to related folders
