+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
 import PageContainer from "@/components/layout/page-container";
-import PageHeader from "@/components/layout/page-header";
+import PageHeader from "@/components/layout/page-header/page-header";
 import ReactQueryProvider from "@/components/providers/react-query";
 import type { Metadata } from "next";
 import { Geologica } from "next/font/google";
