+++ src/components/layout/page-header/page-menu-categories.tsx
@@ -1,9 +1,9 @@
+import { Button } from "@/components/ui/button";
+import { ScrollArea } from "@/components/ui/scroll-area";
 import { cn } from "@/lib/util";
 import { Category } from "@/service/category";
 import Link from "next/link";
 import { HTMLAttributes } from "react";
-import { Button } from "../ui/button";
-import { ScrollArea } from "../ui/scroll-area";
 
 interface PageMenuCategoryProps extends HTMLAttributes<HTMLLIElement> {
   category: Category;
